String Based Approach in the controller
The difference between Model, ModelMap and ModelAndView Objects are now
clear.
