+++ src/main/resources/templates/playingScreen.html
@@ -5,9 +5,13 @@
 <title>Playing</title>
 </head>
 <body th:object="${model}">
+	<h2>Playing....</h2>
+	
+	
 
 	<form th:action="@{/moo/playing}" action="#" method="POST"   >
-		<p th:if="*{message}" th:text="*{message}">asdasd</p>
+		<p th:if="${errorMessage}" th:text="${errorMessage}">Error Message</p>
+
 		<label>Enter your a guess</label> 
 		<input type="number" th:field="*{guess}" /> 
 		<input type="submit"
